// lib/giata-client/giata-types.ts
// Generated by https://quicktype.io

export interface IGiataListPropertiesResponse {
    properties: IProperties
}

export interface IProperties {
    attributes: IPropertiesAttributes
    property: IGiataProperty[]
    more: IMore
}

export interface IPropertiesAttributes {
    'xmlns:xsi': string
    'xmlns:xlink': string
    lastUpdate: string
}

export interface IMore {
    attributes: IMoreAttributes
}

export interface IMoreAttributes {
    'xlink:href': string
}

export interface IGiataProperty {
    attributes: IPropertyAttributes
    name: string
    alternativeNames: IAlternativeNames
    city: ICity
    destination: IDestination
    country: string
    category: number
    ratings: IRatings
    airports: IAirports
    addresses: IAddresses
    phones: IPhones
    emails: IEmails
    urls: IUrls
    geoCodes: IGeoCodes
    propertyCodes: IPropertyCodes
    chains: IChains
    ghgml: IMore
}

export interface IAddresses {
    address: IAddress
}

export interface IAddress {
    addressLine: IAddressLine[]
    street: string
    cityName: string
    country: string
}

export interface IAddressLine {
    text: string
    attributes: IAddressLineAttributes
}

export interface IAddressLineAttributes {
    addressLineNumber: string
}

export interface IAirports {
    airport: IAirport[]
}

export interface IAirport {
    attributes: IAirportAttributes
}

export interface IAirportAttributes {
    iata: string
}

export interface IAlternativeNames {
    alternativeName: IAlternativeName[]
}

export interface IAlternativeName {
    text: string
    attributes: IAlternativeNameAttributes
}

export interface IAlternativeNameAttributes {
    alternativeNameType: string
    effectiveDate?: string
}

export interface IPropertyAttributes {
    giataId: string
    lastUpdate: string
}

export interface IChains {
    chain: IChain
}

export interface IChain {
    attributes: IChainAttributes
}

export interface IChainAttributes {
    chainId: string
    chainName: string
}

export interface ICity {
    text: string
    attributes: ICityAttributes
}

export interface ICityAttributes {
    cityId: string
}

export interface IDestination {
    text: string
    attributes: IDestinationAttributes
}

export interface IDestinationAttributes {
    destinationId: string
}

export interface IEmails {
    email: string
}

export interface IGeoCodes {
    geoCode: IGeoCode
}

export interface IGeoCode {
    attributes: IGeoCodeAttributes
    latitude: number
    longitude: number
}

export interface IGeoCodeAttributes {
    accuracy: string
}

export interface IPhones {
    phone: IPhone[]
}

export interface IPhone {
    text: number
    attributes: IPhoneAttributes
}

export interface IPhoneAttributes {
    tech: string
}

export interface IPropertyCodes {
    provider: IProvider[]
}

export interface IProvider {
    attributes: IProviderAttributes
    code: ICodeElement[] | IPurpleCode
}

export interface IProviderAttributes {
    providerCode: string
    providerType: IProviderType
}

export enum IProviderType {
    Gds = 'gds',
    TourOperator = 'tourOperator'
}

export interface ICodeElement {
    attributes?: ICodeAttributes
    value: IPurpleValue[] | number | string
}

export interface ICodeAttributes {
    status: IStatus
}

export enum IStatus {
    Inactive = 'inactive'
}

export interface IPurpleValue {
    text: number | string
    attributes: IValueAttributes
}

export interface IValueAttributes {
    name: string
}

export interface IPurpleCode {
    value: IFluffyValue[] | number | string
}

export interface IFluffyValue {
    text: string
    attributes: IValueAttributes
}

export interface IRatings {
    rating: IRating
}

export interface IRating {
    attributes: IRatingAttributes
}

export interface IRatingAttributes {
    value: string
    isDefault: string
}

export interface IUrls {
    url: string[]
}